add animations and change statusBar and navigationBar color

+++ data/src/main/java/com/example/data/datasource/dao/WorkDao.kt
@@ -7,7 +7,7 @@ import com.example.data.model.WorkModel
 interface WorkDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun addWork(workModel: WorkModel): Int
+    fun addWork(workModel: WorkModel): Long
 
     @Query("SELECT * FROM work_table")
     fun getAllWorks(): List<WorkModel>
